Fix/multi: views, serializers, models, admin

+++ store/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, get_object_or_404
 from rest_framework.generics import ListAPIView, RetrieveAPIView, CreateAPIView, GenericAPIView, DestroyAPIView
-from .serializers import ProductSerializer, CartSerializer, ClientSerializer, AddToCartSerializer, DeleteCartItemSerializer, OrderSerializer, CouponCodeSerializer
+from .serializers import ProductSerializer, CartSerializer, ClientSerializer, CartItemQuantitySerializer, OrderSerializer, CouponCodeSerializer
 from .models import Product, Cart, Client, CartItem
 from rest_framework import viewsets
 from rest_framework import permissions, status
@@ -20,17 +20,9 @@ class ProductsStoreView(ListAPIView):
     serializer_class = ProductSerializer    
     queryset = Product.objects.all()
 
-class CartViewSet(viewsets.ModelViewSet): # --> alterar essa viewset para uma generic view
-    permission_classes = [permissions.IsAuthenticated]
-    serializer_class = CartSerializer
-    queryset = Cart.objects.all()
-
     def get_queryset(self):
-        queryset = Cart.objects.filter(user=self.request.user)
+        queryset = Product.objects.filter(active=True)
         return queryset
-    
-    def perform_create(self, serializer):
-        serializer.save(user=self.request.user)
     
 class ProductDetail(RetrieveAPIView):
     permission_classes = [permissions.IsAuthenticated]
@@ -42,42 +34,51 @@ class AddToCart(APIView):
     permission_classes = [permissions.IsAuthenticated]
         
     def post(self, request, pk, *args, **kwargs,):
-        serializer = AddToCartSerializer(data=request.data)
+        serializer = CartItemQuantitySerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
         product = Product.objects.get(id=pk)
         quantity = serializer.validated_data.get('product_quantity')
 
-        cart, _= Cart.objects.get_or_create(user=request.user)
+        cart, _= Cart.get_cart(user=request.user)
 
         cart_item, created = CartItem.objects.get_or_create(cart=cart, product=product)
-        if not created:
-            cart_item.quantity += quantity
-        else:
-            cart_item.quantity = quantity
+
+        total_quantity = quantity if created else quantity + cart_item.quantity
+
+        if not product.active:
+            return Response({"error": "This product isn't active"}, status=status.HTTP_404_NOT_FOUND)
+        if product.stock < total_quantity:
+            return Response({"error": "The quantity to add exceeds available stock"},status=status.HTTP_400_BAD_REQUEST)
+        
+        cart_item.quantity = total_quantity
         cart_item.save()
 
-        return Response({'detail': 'Product Added To Cart', 'cart_subtotal': cart_item.subtotal()}, status=status.HTTP_200_OK)
+        return Response({'detail': 'Product Added To Cart', 'cart_subtotal': cart_item.subtotal(), 'cart_total': cart.total(), 'total_items': cart.total_items()}, status=status.HTTP_200_OK)
     
 class DeleteCartItem(APIView):
     permission_classes =[permissions.IsAuthenticated]
 
     def delete(self, request, pk, *args, **kwargs):
-        serializer = DeleteCartItemSerializer(data=request.data)
+        serializer = CartItemQuantitySerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
-        cart = Cart.objects.get(user=request.user)
-        cart_item = cart.items.filter(product__id=pk).first()
+        cart, _ = Cart.get_cart(user=request.user)
+        if not cart:
+            return Response({"error": "You don't have an active cart"},status=status.HTTP_404_NOT_FOUND)
+        
+        cart_item = cart.items.filter(product__id=pk, product__active=True).first()
         qtd_to_remove = serializer.validated_data.get('product_quantity')
 
         if cart_item:
             cart_item.quantity -= qtd_to_remove
             if cart_item.quantity <= 0:
                 cart_item.delete()
-                return Response({'detail': 'The product was completely removed from your cart', 'cart_subtotal': cart_item.subtotal()},status=status.HTTP_200_OK)
+                return Response({'detail':  'The product was completely removed from your cart', 'cart_total': cart.total()},status=status.HTTP_200_OK)
             else:
                 cart_item.save()
-                return Response({'detail': 'The product quantity has been updated in your cart', 'cart_subtotal': cart_item.subtotal()},status=status.HTTP_200_OK)
+                subtotal = cart_item.subtotal()
+                return Response({'detail': 'The product quantity has been updated in your cart', 'cart_subtotal': subtotal, 'cart_total': cart.total(), 'total_items': cart.total_items()},status=status.HTTP_200_OK)
         
         return Response({"error": "The product doesn't exist in your cart"}, status=status.HTTP_404_NOT_FOUND)
     
@@ -87,7 +88,7 @@ class Continue_Payment(APIView): #Mudar esse nome depois
     permission_classes = [permissions.IsAuthenticated]
     
     def get(self, request, *args, **kwargs):
-        cart, _ = Cart.objects.get_or_create(user=request.user)
+        cart, _ = Cart.get_cart(user=request.user)
 
         query_serializer = CouponCodeSerializer(data=request.query_params)
         query_serializer.is_valid(raise_exception=True)
@@ -95,7 +96,7 @@ class Continue_Payment(APIView): #Mudar esse nome depois
         coupon_code = query_serializer.validated_data.get('coupon_code')
 
         if not cart.items.exists():
-            return Response({"error':'You don't have items in your cart"}, status=status.HTTP_404_NOT_FOUND)
+            return Response({"error": "You don't have items in your cart"}, status=status.HTTP_404_NOT_FOUND)
         serializer = CartSerializer(cart, context={'coupon_code': coupon_code})
         return Response(serializer.data)
     
@@ -104,7 +105,7 @@ class CreateOrder(APIView):
 
     def post(self, request, *args, **kwargs):
 
-        cart, _ = Cart.objects.get_or_create(user=request.user, finalized=False)
+        cart, _ = Cart.get_cart(user=request.user)
 
         if not cart.items.exists():
             return Response({"error": "You don't have items in your cart"}, status=status.HTTP_404_NOT_FOUND)
